requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
sqlalchemy==2.0.23
pandas==2.1.4
networkx==3.2.1
groq==0.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
